updating to `next.415` sveltekit `V` for `bull-workers`
