+++ src/routes/api/bull/featured_betting_sites/+server.ts
@@ -6,6 +6,7 @@ import { GET_TRANSLATIONS_DATA_FEATURED_BETTING_SITES } from '$lib/graphql/featu
 import { GET_HREFLANG_DATA } from '$lib/graphql/query'
 import { performance } from 'perf_hooks';
 import Bull from 'bull';
+import { error, json } from '@sveltejs/kit';
 
 import type { 
   All_SportBook_Details_Data, 
@@ -41,7 +42,7 @@ let logs = []
 //  [MAIN] ENDPOINT METHOD
 // ~~~~~~~~~~~~~~~~~~~~~~~~
 
-export async function post(): Promise < unknown > {
+export async function POST(): Promise < unknown > {
 
   // [🐛] debug
   if (dev) console.log(`
@@ -56,13 +57,9 @@ export async function post(): Promise < unknown > {
     job_id: ${job.id}
   `)
 
-  return {
-    status: 200,
-    body: { 
-      job_id: job.id
-    }
-  }
-
+  return json({
+    job_id: job.id
+  })
 }
 
 // ~~~~~~~~~~~~~~~~~~~~~~~~
