+++ src/routes/api/bull/league_list/+server.ts
@@ -5,6 +5,7 @@ import { initGrapQLClient } from '$lib/graphql/init_graphQL'
 import fs from 'fs';
 import { performance } from 'perf_hooks';
 import Bull from 'bull';
+import { error, json } from '@sveltejs/kit';
 
 import type { 
   BETARENA_HASURA_league_list_query,
@@ -53,7 +54,7 @@ let logs = []
 //  [MAIN] ENDPOINT METHOD
 // ~~~~~~~~~~~~~~~~~~~~~~~~
 
-export async function post(): Promise < unknown > {
+export async function POST(): Promise < unknown > {
 
   // [🐛] debug
   if (dev) console.log(`
@@ -68,12 +69,9 @@ export async function post(): Promise < unknown > {
     job_id: ${job.id}
   `)
 
-  return {
-    status: 200,
-    body: { 
-      job_id: job.id
-    }
-  }
+return json({
+    job_id: job.id
+  })
 
 }
