+++ src/routes/api/bull/leagues_table/+server.ts
@@ -4,6 +4,7 @@ import { initGrapQLClient } from '$lib/graphql/init_graphQL'
 import fs from 'fs';
 import { performance } from 'perf_hooks';
 import Bull from 'bull';
+import { error, json } from '@sveltejs/kit';
 
 import { 
   REDIS_CACHE_LEAGUES_TABLE_DATA_1, 
@@ -52,7 +53,7 @@ let logs = []
 //  [MAIN] ENDPOINT METHOD
 // ~~~~~~~~~~~~~~~~~~~~~~~~
 
-export async function post(): Promise < unknown > {
+export async function POST(): Promise < unknown > {
 
   // [🐛] debug
   if (dev) console.log(`
@@ -67,12 +68,9 @@ export async function post(): Promise < unknown > {
     job_id: ${job.id}
   `)
 
-  return {
-    status: 200,
-    body: { 
-      job_id: job.id
-    }
-  }
+return json({
+    job_id: job.id
+  })
 }
 
 // ~~~~~~~~~~~~~~~~~~~~~~~~
