+++ src/routes/api/bull/navbar/+server.ts
@@ -4,6 +4,7 @@ import { initGrapQLClient } from '$lib/graphql/init_graphQL';
 import { GET_NAVBAR_DATA } from '$lib/graphql/header/query';
 import { performance } from 'perf_hooks';
 import Bull from 'bull';
+import { error, json } from '@sveltejs/kit';
 
 import type { 
   Cache_Single_Lang_Header_Translation_Response, 
@@ -38,7 +39,7 @@ let logs = []
 //  [MAIN] ENDPOINT METHOD
 // ~~~~~~~~~~~~~~~~~~~~~~~~
 
-export async function post(): Promise < unknown > {
+export async function POST(): Promise < unknown > {
 
   // [🐛] debug
   if (dev) console.log(`
@@ -53,12 +54,9 @@ export async function post(): Promise < unknown > {
     job_id: ${job.id}
   `)
 
-  return {
-    status: 200,
-    body: { 
-      job_id: job.id
-    }
-  }
+return json({
+    job_id: job.id
+  })
 }
 
 // ~~~~~~~~~~~~~~~~~~~~~~~~
