+++ src/routes/api/bull/pages_and_seo/+server.ts
@@ -5,6 +5,7 @@ import { removeDiacritics } from '$lib/utils/languages'
 import fs from 'fs';
 import { performance } from 'perf_hooks';
 import Bull from 'bull';
+import { error, json } from '@sveltejs/kit';
 
 import { createRequire } from 'module';
 const require = createRequire(import.meta.url);
@@ -51,7 +52,7 @@ let logs = []
 //  [MAIN] ENDPOINT METHOD
 // ~~~~~~~~~~~~~~~~~~~~~~~~
 
-export async function post(): Promise < unknown > {
+export async function POST(): Promise < unknown > {
 
   // [🐛] debug
   if (dev) console.log(`
@@ -66,12 +67,9 @@ export async function post(): Promise < unknown > {
     job_id: ${job.id}
   `)
 
-  return {
-    status: 200,
-    body: { 
-      job_id: job.id
-    }
-  }
+return json({
+    job_id: job.id
+  })
 }
 
 // ~~~~~~~~~~~~~~~~~~~~~~~~
