+++ src/routes/api/bull/seo_block/+server.ts
@@ -1,14 +1,15 @@
 import { dev } from '$app/env'
+
 import redis from "$lib/redis/init"
 import { initGrapQLClient } from '$lib/graphql/init_graphQL';
-import { GET_FOOTER_DATA } from '$lib/graphql/query';
+import { GET_HREFLANG_DATA } from '$lib/graphql/query';
+import { GET_COMPLETE_PAGES_AND_SEO_DATA } from '$lib/graphql/pages_and_seo/query';
 import { performance } from 'perf_hooks';
 import Bull from 'bull';
+import { error, json } from '@sveltejs/kit';
 
-import type { 
-  Cache_Single_Lang_Footer_Translation_Response, 
-  Hasura_Footer_Translation_Response 
-} from '$lib/models/footer/types';
+import type { Cache_Single_Homepage_SEO_Block_Translation_Response } from '$lib/models/seo_block/types';
+import type { Hasura_Complete_Pages_SEO } from '$lib/models/pages_and_seo/types';
 
 // ~~~~~~~~~~~~~~~~~~~~~~~~
 // [❗] BULL CRITICAL
@@ -19,8 +20,8 @@ const settings = {
   guardInterval: 5000, // Poll interval for delayed jobs and added jobs.
   drainDelay: 300 // A timeout for when the queue is in drained state (empty waiting for jobs).
 }
-const cacheQueueFooter = new Bull (
-  'cacheQueueFooter', 
+const cacheQueueSeoBlock = new Bull (
+  'cacheQueueSeoBlock', 
   { 
     redis: { 
       port: import.meta.env.VITE_REDIS_BULL_ENDPOINT.toString(), 
@@ -31,14 +32,14 @@ const cacheQueueFooter = new Bull (
     settings: settings
   }
 );
-const cacheTarget = "REDIS CACHE | footer"
+const cacheTarget = "REDIS CACHE | seo_block"
 let logs = []
 
 // ~~~~~~~~~~~~~~~~~~~~~~~~
 //  [MAIN] ENDPOINT METHOD
 // ~~~~~~~~~~~~~~~~~~~~~~~~
 
-export async function post(): Promise < unknown > {
+export async function POST(): Promise < unknown > {
 
   // [🐛] debug
   if (dev) console.log(`
@@ -47,38 +48,36 @@ export async function post(): Promise < unknown > {
   `);
 
   // [ℹ] producers [JOBS]
-  const job = await cacheQueueFooter.add({});
+  const job = await cacheQueueSeoBlock.add({}, { timeout: 180000 });
 
   console.log(`
     job_id: ${job.id}
   `)
 
-  return {
-    status: 200,
-    body: { 
-      job_id: job.id
-    }
-  }
+return json({
+    job_id: job.id
+  })
+
 }
 
 // ~~~~~~~~~~~~~~~~~~~~~~~~
 //  [MAIN] CACHING METHODS
 // ~~~~~~~~~~~~~~~~~~~~~~~~
 
-async function cacheFooter(lang: string, json_cache: Cache_Single_Lang_Footer_Translation_Response) {
+async function cacheSEOblock(lang: string, json_cache: Cache_Single_Homepage_SEO_Block_Translation_Response) {
   try {
     //[ℹ] persist redis (cache)
-    await redis.hset('footer_t', lang, JSON.stringify(json_cache));
+    await redis.hset('seo_block_t', lang, JSON.stringify(json_cache));
     // [🐛] debug
-    if (dev) console.debug('✅ footer-data cached')
+    if (dev) console.debug('✅ navbar-data cached')
   } 
   catch (e) {
-    console.error("❌ unable to cache", 'footer-data', e);
+    console.error("❌ unable to cache", 'navbar-data', e);
   }
 }
 
-async function deleteCacheFooter() {
-  await redis.del('footer_t')
+async function deleteSEOblock() {
+  await redis.del('seo_block_t')
   return
 }
 
@@ -86,7 +85,7 @@ async function deleteCacheFooter() {
 //  [MAIN] BULL WORKERS 
 // ~~~~~~~~~~~~~~~~~~~~~~~~
 
-cacheQueueFooter.process (async (job, done) => {
+cacheQueueSeoBlock.process (async (job, done) => {
   // console.log(job.data.argumentList);
 
   logs = []
@@ -120,8 +119,8 @@ cacheQueueFooter.process (async (job, done) => {
 
 async function main() {
 
-  // [ℹ] get HASURA-DB response;
-	const response: Hasura_Footer_Translation_Response = await initGrapQLClient().request(GET_FOOTER_DATA);
+  // [ℹ] get KEY platform translations
+  const response = await initGrapQLClient().request(GET_HREFLANG_DATA)
 
   // [ℹ] get-all-exisitng-lang-translations;
   const langArray: string [] = response.scores_hreflang_dev
@@ -134,23 +133,26 @@ async function main() {
   // [🐛] debug
   if (dev) console.debug("langArray", langArray)
 
-  const finalCacheObj: Cache_Single_Lang_Footer_Translation_Response = {
+  const finalCacheObj: Cache_Single_Homepage_SEO_Block_Translation_Response = {
     lang: undefined,
-    scores_footer_translations_dev: undefined,
-    scores_footer_links_dev: undefined,
+    html: undefined,
+    title: undefined,
   }
 
-  // deleteCacheFooter()
+  // deleteSEOblock()
+
+  const response_data: Hasura_Complete_Pages_SEO = await initGrapQLClient().request(GET_COMPLETE_PAGES_AND_SEO_DATA)
 
   // [ℹ] for-each available translation:
   for (const lang_ of langArray) {
     
     finalCacheObj.lang = lang_;
-    finalCacheObj.scores_footer_translations_dev = response.scores_footer_translations_dev.find(( { lang } ) => lang_ === lang);
-    finalCacheObj.scores_footer_links_dev = response.scores_footer_links_dev.find(( { lang } ) => lang_ === lang);
+
+    finalCacheObj.html = response_data.scores_seo_block_homepage_dev.find(( { lang } ) => lang_ === lang).html;
+    finalCacheObj.title = response_data.scores_seo_block_homepage_dev.find(( { lang } ) => lang_ === lang).title;
 
     // [ℹ] persist-cache-response;
-    await cacheFooter(lang_, finalCacheObj);
+    await cacheSEOblock(lang_, finalCacheObj);
   }
 
   return;
