+++ src/routes/api/bull/tournaments_standings/+server.ts
@@ -5,6 +5,7 @@ import { GET_LEAGUE_INFO_FULL_DATA } from '$lib/graphql/tournaments/query';
 import fs from 'fs';
 import Bull from 'bull';
 import { performance } from 'perf_hooks';
+import { error, json } from '@sveltejs/kit';
 
 import type { 
   BETARENA_HASURA_tournament_standings_query,
@@ -45,7 +46,7 @@ let logs = []
 //  [MAIN] ENDPOINT METHOD
 // ~~~~~~~~~~~~~~~~~~~~~~~~
 
-export async function post(): Promise < unknown > {
+export async function POST(): Promise < unknown > {
   
   // [🐛] debug
   if (dev) console.log(`
@@ -60,12 +61,9 @@ export async function post(): Promise < unknown > {
     job_id: ${job.id}
   `)
 
-  return {
-    status: 200,
-    body: { 
-      job_id: job.id
-    }
-  }
+return json({
+    job_id: job.id
+  })
 }
 
 // ~~~~~~~~~~~~~~~~~~~~~~~~
