+++ src/routes/api/bull/tournaments_top_players/+server.ts
@@ -7,6 +7,7 @@ import {
 import fs from 'fs';
 import { performance } from 'perf_hooks';
 import Bull from 'bull';
+import { error, json } from '@sveltejs/kit';
 
 import type { 
   BETARENA_HASURA_top_players_query, 
@@ -57,7 +58,7 @@ let logs = []
 //  [MAIN] ENDPOINT METHOD
 // ~~~~~~~~~~~~~~~~~~~~~~~~
 
-export async function post(): Promise < unknown > {
+export async function POST(): Promise < unknown > {
 
   // [ℹ] dev / local environment
   if (dev) {
